# nana_web/openapi.json
{
  "components": {
    "schemas": {
...
      "ContractCatalogInfo": {
        "additionalProperties": false,
        "properties": {
          "command_names": {
            "items": {
              "type": "string"
            },
            "title": "Command Names",
            "type": "array"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "title": "Event Types",
            "type": "array"
          },
          "locator_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Locator Kinds",
            "type": "array"
          },
          "relation_types": {
            "items": {
              "type": "string"
            },
            "title": "Relation Types",
            "type": "array"
          },
          "schema_hash": {
            "title": "Schema Hash",
            "type": "string"
          },
          "schema_names": {
            "items": {
              "type": "string"
            },
            "title": "Schema Names",
            "type": "array"
          }
        },
        "required": [
          "schema_hash",
          "schema_names",
          "command_names",
          "event_types",
          "locator_kinds",
          "relation_types"
        ],
        "title": "ContractCatalogInfo",
        "type": "object"
      },
...
      "ExportSelectionInfo": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "provenance": {
            "enum": [
              "interactive_user",
              "test_harness"
            ],
            "title": "Provenance",
            "type": "string"
          },
          "selection_id": {
            "title": "Selection Id",
            "type": "string"
          }
        },
        "required": [
          "selection_id",
          "label",
          "expires_at",
          "provenance"
        ],
        "title": "ExportSelectionInfo",
        "type": "object"
      },
...
      "RuntimeHandshakeResponse": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "app_version": {
            "title": "App Version",
            "type": "string"
          },
          "enabled_mutations": {
            "items": {
              "type": "string"
            },
            "title": "Enabled Mutations",
            "type": "array"
          },
          "execution_enabled": {
            "default": false,
            "title": "Execution Enabled",
            "type": "boolean"
          },
          "export_selections": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExportSelectionInfo"
            },
            "title": "Export Selections",
            "type": "array"
          },
          "external_effects_enabled": {
            "default": false,
            "title": "External Effects Enabled",
            "type": "boolean"
          },
          "mode": {
            "const": "development_runtime",
            "default": "development_runtime",
            "title": "Mode",
            "type": "string"
          },
          "mutations_enabled": {
            "const": true,
            "default": true,
            "title": "Mutations Enabled",
            "type": "boolean"
          },
          "schema_read_ceiling": {
            "title": "Schema Read Ceiling",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "app_version",
          "api_version",
          "schema_version",
          "schema_read_ceiling",
          "enabled_mutations"
        ],
        "title": "RuntimeHandshakeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "LocalSessionBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nana Sidecar Runtime",
    "version": "0.3.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/contracts": {
      "get": {
        "operationId": "contract_catalog_api_v1_contracts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCatalogInfo"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "LocalSessionBearer": []
          }
        ],
        "summary": "Contract Catalog",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/v1/handshake": {
      "get": {
        "operationId": "handshake_api_v1_handshake_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeHandshakeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "LocalSessionBearer": []
          }
        ],
        "summary": "Handshake",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}